Skip the windows ASLR test when CGO_ENABLED=0
