+++ src/cmd/link/internal/ld/ld_test.go
@@ -160,22 +160,6 @@ func TestPPC64LargeTextSectionSplitting(t *testing.T) {
 	}
 }
 
-func TestWindowsBuildmodeCSharedASLR(t *testing.T) {
-	platform := fmt.Sprintf("%s/%s", runtime.GOOS, runtime.GOARCH)
-	switch platform {
-	case "windows/amd64", "windows/386":
-	default:
-		t.Skip("skipping windows amd64/386 only test")
-	}
-
-	t.Run("aslr", func(t *testing.T) {
-		testWindowsBuildmodeCSharedASLR(t, true)
-	})
-	t.Run("no-aslr", func(t *testing.T) {
-		testWindowsBuildmodeCSharedASLR(t, false)
-	})
-}
-
 func testWindowsBuildmodeCSharedASLR(t *testing.T, useASLR bool) {
 	t.Parallel()
 	testenv.MustHaveGoBuild(t)
